+++ problem_space/cell/cell.h
@@ -12,10 +12,10 @@
 #include "neighbor.h"
 
 struct cell{
-    float r_a; //hamiltonian of SOS method
-    float r_ei; //hamiltonian of DG method
+    double r_a; //hamiltonian of SOS method
+    double r_ei; //hamiltonian of DG method
     int h;// cell height
-    float cumulative_r_ei_index;
+    double cumulative_r_ei_index;
 };
